% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_p.r
\name{brownsMethod}
\alias{brownsMethod}
\title{Merge p-values using Brown's method}
\usage{
brownsMethod(p.values, data.matrix = NULL, cov.matrix = NULL)
}
\arguments{
\item{p.values}{A vector of m p-values}

\item{data.matrix}{An m x n matrix representing m tests and n samples}

\item{cov.matrix}{A pre-calculated covariance matrix of data.matrix. More
efficient when making many calls with the same data.matrix.
Only one of data.matrix and cov.matrix must be given. If both are supplied,
data.matrix is ignored}
}
\value{
a p-value
}
\description{
Based on the R package EmpiricalBrownsMethod
https://github.com/IlyaLab/CombiningDependentPvaluesUsingEBM/blob/master/R/EmpiricalBrownsMethod/R/ebm.R
Only significant differences are the removal of extra_info and allowing a
pre-calculated covariance matrix
}
