% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmt.r
\name{GMT}
\alias{GMT}
\alias{gmt}
\alias{read.GMT}
\alias{write.GMT}
\title{Read and Write GMT files}
\format{A named list of terms, where each term is a list with the items:
\describe{
    \item{id}{The term id}
    \item{name}{The full name of the term}
    \item{genes}{A character vector of genes annotated to this term}
  }}
\usage{
read.GMT(filename)

write.GMT(gmt, filename)
}
\arguments{
\item{filename}{Location of the gmt file}

\item{gmt}{a GMT object}
}
\value{
\code{read.GMT} returns a GMT object. \cr
\code{write.gmt} returns NULL.
}
\description{
Functions to read and Write Gene Matrix Transposed (GMT) files. A GMT file
describes gene sets, such as pathways. GMT files are tab delimeted and each
row contains a term id, a term name, and all genes annotated to the term.
}
