% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tests.r
\name{hypergeometric}
\alias{hypergeometric}
\title{Hypergeometric Test}
\usage{
hypergeometric(genelist, complement, annotations)
}
\arguments{
\item{genelist}{character vector of gene names. List of differentially 
expressed genes being tested for enrichment}

\item{complement}{character vector of gene names. All other genes not being
tested for enrichment. Ie \code{setdiff(background, genelist)}}

\item{annotations}{character vector of gene names. List of genes annotated
to the term being tested.}
}
\value{
a p-value
}
\description{
Perform a Hypergeometric, aka Fisher's exact test, on a list of significant
  genes and annotation genes
}
