% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmt.r
\name{makeBackground}
\alias{makeBackground}
\title{Make a background list of genes (i.e., the statistical universe) based on all the terms (gene sets, pathways) considered.}
\usage{
makeBackground(gmt)
}
\arguments{
\item{gmt}{A \link{GMT} object.}
}
\value{
A character vector containing all genes in GMT.
}
\description{
Returns A character vector of all genes in a GMT object.
}
\examples{
\dontrun{
  makeBackground(gmt)
}
}
