% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmt.r
\name{readGMT}
\alias{readGMT}
\title{Read a gmt file}
\usage{
readGMT(filename)
}
\arguments{
\item{filename}{a character vector containing the location of the gmt file}
}
\value{
a gmt object, which is a lest of terms where each term is a list with the items:
  \itemize{
    \item{"id"}{The term id}
    \item{"name"}{The full name of the term}
    \item{"genes"}{A character vector of genes annotated to this term}
  }
}
\description{
Returns a gmt object read in from a file
}
