% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmt.r
\name{write.GMT}
\alias{write.GMT}
\title{Write a gmt to file}
\usage{
write.GMT(gmt, filename)
}
\arguments{
\item{gmt}{a gmt object}

\item{filename}{Where to write the gmt file to}
}
\description{
Write a gmt to file
}
