% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpea.r
\name{columnContribution}
\alias{columnContribution}
\title{Get contribution of each column to each term}
\usage{
columnContribution(scores, gmt, background, significant, cutoff,
  correction.method)
}
\arguments{
\item{scores}{A numerical matrix of p-values where each row is a gene and
each column is a test. Rownames should be the genes and colnames the names
of the tests. All values must be 0<=p<=1 with missing values removed or
converted to 1}

\item{gmt}{A GMT object to be used for enrichment analysis. If a filename, a
GMT object will be read from the file}

\item{background}{A character vector of gene names to be used as a
statistical background. By default, the background is all genes that appear
in \code{gmt}}

\item{significant}{A number in [0,1] denoting the maximum p-value for a
pathway to be considered significantly enriched.}

\item{cutoff}{A maximum p-value for agene to be used for enrichment analysis.
Any genes with \code{p.val > significant} will be discarded before testing}

\item{correction.method}{Method to correct p-values. See
\code{\link[stats]{p.adjust}} for details}
}
\value{
a data.table of terms. The first two columns contain the term id and
  name. The rest of the columns denoting whether the term is found to be
  significant if using only that test in analysis
}
\description{
Get contribution of each column to each term
}
\examples{
\dontrun{
  dat <- as.matrix(read.table('path/to/data.txt', header=TRUE, row.names='Gene'))
  dat[is.na(dat)] <- 1
  gmt <- read.GMT('path/to/gmt.gmt')
  columnContribution(dat, gmt, makeBackground(gmt), 0.05, 0.1, 'fdr')
}
}
