% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tests.r
\name{hypergeometric}
\alias{hypergeometric}
\title{Hypergemoetric Test}
\usage{
hypergeometric(counts)
}
\arguments{
\item{counts}{a 2x2 numerical matrix representing a contingency table}
}
\value{
a p-value
}
\description{
Perform a Hypergeometric test, aka Fisher's exact test, on a 2x2 contingency
table with a 'greater' alternative hypothesis. That is, find the
probability that counts[1, 1] or more genes would be found in annotations,
assuming the null hypothesis.
}
\examples{
hypergeometric(matrix(data=c(1, 8, 14, 836), nrow=2))

}
