% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmt.r
\name{makeBackground}
\alias{makeBackground}
\title{Make a background list of genes}
\usage{
makeBackground(gmt)
}
\arguments{
\item{gmt}{a \link{GMT} object}
}
\value{
a character vector containing all genes in GMT
}
\description{
Returns a character vector of all genes in a GMT object
}
\examples{
\dontrun{
  makeBackground(gmt)
}
}
