% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmt.r
\name{makeBackground}
\alias{makeBackground}
\title{Make a background list of genes (i.e., the statistical universe) based on all the terms (gene sets, pathways) considered.}
\usage{
makeBackground(gmt)
}
\arguments{
\item{gmt}{A \link{GMT} object.}
}
\value{
A character vector containing all genes in GMT.
}
\description{
Returns A character vector of all genes in a GMT object.
}
\examples{
  fname_GMT <- system.file("extdata", "hsapiens_REAC_subset.gmt", package = "ActivePathways")
  gmt <- read.GMT(fname_GMT)
  makeBackground(gmt)[1:10]
}
