% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_p.r
\name{merge_p_values}
\alias{merge_p_values}
\title{Merge a list or matrix of p-values}
\usage{
merge_p_values(scores, method = c("Fisher", "Brown", "logitp", "meanp",
  "sump", "sumz", "sumlog"))
}
\arguments{
\item{scores}{Either a list of p-values or a matrix where each column is a test}

\item{method}{Method to merge p-valeus. See metap documentation for more info}
}
\value{
If scores is a list, returns a number. If scores is a matrix, returns
  a named list of p-values merged by row
}
\description{
Merge a list or matrix of p-values
}
