% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpea.r
\name{mpea}
\alias{mpea}
\title{mpea}
\usage{
mpea(scores, gmt, cutoff = 0.1, significant = 0.05, return.all = FALSE,
  merge.method = c("Fisher", "Brown", "logitp", "meanp", "sump", "sumz",
  "sumlog"), correction.method = c("holm", "fdr", "hochberg", "hommel",
  "bonferroni", "BH", "BY"), background = makeBackground(gmt),
  contribution = TRUE, cytoscape.filenames = NULL)
}
\arguments{
\item{scores}{A numerical matrix of p-values where each row is a gene and
each column is a test. Rownames should be the genes and colnames the names
of the tests. All values must be 0<=p<=1 with missing values removed or
converted to 1}

\item{gmt}{A GMT object to be used for enrichment analysis. If a filename, a
GMT object will be read from the file}

\item{cutoff}{A maximum p-value for agene to be used for enrichment analysis.
Any genes with \code{p.val > significant} will be discarded before testing}

\item{significant}{A number in [0,1] denoting the maximum p-value for a
pathway to be considered significantly enriched.}

\item{return.all}{Whether to return results for all terms or only significant
terms}

\item{merge.method}{Method to merge p-values.}

\item{correction.method}{Method to correct p-values. See
\code{\link[stats]{p.adjust}} for details}

\item{background}{A character vector of gene names to be used as a
statistical background. By default, the background is all genes that appear
in \code{gmt}}

\item{contribution}{Evaluate contribution of individual columns to a term's
significance. See section Column Contribution}

\item{cytoscape.filenames}{a vector of 2 or 3 filenames denoting where
information for Cytoscape will be written to (see section Cytoscape). If
NULL, do not write any files.}
}
\value{
A data.table of terms containing the following columns:
  \describe{
    \item{term.id}{The id of the term}
    \item{term.name}{The full name of the term}
    \item{p.val}{The associated p-value}
    \item{term.size}{The number of genes annotated to the term}
    \item{overlap}{A character vector of the genes that overlap between the
    term and the query}
    \item{...}{If \code{contribution == TRUE}, A column for each test in
      \code{scores} denoting whether the term is found to be significant if
      using only that test in analysis}
  }
  If \code{return.all == FALSE} then only terms with
    \code{p.val <= significant} will be returned.
}
\description{
mpea
}
\section{Cytoscape}{

  If \code{cytoscape.filenames} is supplied, \code{mpea} will write three
  files that can be used to build an network using Cytoscape and the
  EnrichmentMap app. The three fies written are:
  \describe{
    \item{cytoscape.filenames[1]}{A list of significant terms and the
    associated p-value. Only terms with \code{p.val <= significant} are
    written to this file}
    \item{cytoscape.filenames[2]}{A matrix of column contributions to each
    term. A 1 indicates that that term is significant using only that
    column to test for enrichment analysis}
    \item{cytoscape.filenames[3]}{A Shortened version of the supplied gmt
    file, containing only the terms in \code{cytoscape.filenames[1]}}
  }
  If \code{contribution == FALSE} the matrix of column contributions will not
  be written. Only 2 file names need to be supplied, and if three are given
  the second will be ignored

  Use: Create an enrichment map in Cytoscape with the file of terms
  (cytoscape.filenames[1]) and the shortened gmt file
  (cytoscape.filenames[3]). Upload (File > import > table > file) the
  subgroups file (cytoscape.filenames[2]) and **ensure the `contributes`
  column is a list data type. Under style set image/Chart1 to use the column
  `instruct` and the passthrough mapping type.
}

\section{Merging p-Values}{

In order to obtain a single score for each gene, the p-values in \code{scores}
are merged row-wise. There are multiple methods available that can be used
to obtain this merged score. They are:
\describe{
 \item{Fisher or sumlog}{Fisher's method assumes p-values are uniformly
 distributed and performs a chi-squared test on the statistic sum(-2 log(p).
 This method is most appropriate when the columns in \code{scores} are
 independent.}
 \item{Brown}{Brown's method extends Fisher's method by accounting for the
 covariance in the columns of \code{scores}. It is more appropriate when the
 tests of significance used to create the columns in \code{scores} are not
 necessarily independent.}
}
\item{logitp}{}
\item{meanp}{}
\item{sump}{}
\item{sumz}{}
}

\section{Adjusting p-Values}{

}

\section{Column Contribution}{

}

\examples{
\dontrun{
    dat <- as.matrix(read.table('path/to/data.txt', header=TRUE, row.names='Gene'))
    dat[is.na(dat)] <- 1
    gmt <- read.GMT('path/to/gmt.gmt')
    mpea(dat, gmt, return.all=TRUE,
         cytoscape.filenames=c('terms.txt', 'groups.txt', 'abridged.gmt'))
}

}
