% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpea.r
\name{prepareCytoscape}
\alias{prepareCytoscape}
\title{Prepare files for building an Enrichment Map in Cytoscape}
\usage{
prepareCytoscape(terms, gmt, filenames, contribution)
}
\arguments{
\item{terms}{a data.table of terms, the output of mpea()}

\item{gmt}{A GMT object to be used for enrichment analysis. If a filename,
a GMT object will be created from the file}

\item{contribution}{Evaluate contribution of individual columns to a term's
significance. See section Column Contribution}
}
\value{
None
}
\description{
This function writes three files that are used to build an network using
Cytoscape and the EnrichmentMap app.
  The three fies written are:
  \describe{
    \item{cytoscape.filenames[1]}{A list of significant terms and the
      associated p-value. Only terms with \code{p.val <= significant} are
      written to this file}
    \item{cytoscape.filenames[2]}{A matrix of column contributions to each
      term. A \code{1} indicates that that term is significant using only
      that column to test for enrichment analysis}
    \item{cytoscape.filenames[3]}{A Shortened version of the supplied gmt
      file, containing only the terms in \code{cytoscape.filenames[1]}}
  }
}
