% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoscape.r
\name{prepareCytoscape}
\alias{prepareCytoscape}
\title{Prepare files for building an Enrichment Map in Cytoscape}
\usage{
prepareCytoscape(terms, gmt, filenames, col.significance)
}
\arguments{
\item{terms}{a data.table with columns 'term.id', 'term.name', 'p.val'}

\item{gmt}{an abridged gmt object containing only the pathways that were
found to be significant}

\item{filenames}{vector of 2 or 3 filesnames denoting where to write fles to}

\item{col.significance}{a data.table with a column 'term.id' and a column
for each test indicating whether a pathways is signficiant (TRUE) or not
(FALSE) when considering only that column. If contribution==TRUE, use
col.significance=NULL and this will be skipped}
}
\value{
None
}
\description{
This function writes three files that are used to build an network using
Cytoscape and the EnrichmentMap app.
  The three fies written are:
  \describe{
    \item{cytoscape.filenames[1]}{A list of significant terms and the
      associated p-value. Only terms with \code{p.val <= significant} are
      written to this file}
    \item{cytoscape.filenames[2]}{A matrix of column contributions to each
      term. A \code{1} indicates that that term is significant using only
      that column to test for enrichment analysis}. Left out if
      contribution==FALSE and therefore col.significance==NULL
    \item{cytoscape.filenames[3]}{A Shortened version of the supplied gmt
      file, containing only the terms in \code{cytoscape.filenames[1]}}
  }
}
